fix(restock): urgency_score 改 float（真实数据浮点击穿 strict schema 致 500）
真实 list_sku_summary() 的 urgency_score 是浮点（如 69.5「次紧迫」），
小数位有业务语义：供应商概览按 max urgency 排序露出次紧迫供应商，
旧页 restock.js:152 直接显示原值。schema 原写 int|None → 15269 行
全挂 int_from_float → /api/restock/items 500。本地双栈 + prod 数据
浏览器验收时暴露（合成测试用整数 72 漏过）。

+++ app/schemas_api.py
@@ -510,7 +510,7 @@ class RestockItem(BaseModel):
     restock_qty_p98: int | None
     restock_source: str | None
     last_purchase_qty: int | None
-    urgency_score: int | None
+    urgency_score: float | None  # 真实数据为浮点（69.5「次紧迫」），小数位有排序/显示语义
     stockout_zero_weeks_last8: int
